% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExhaustiveSearch-package.R
\docType{package}
\name{ExhaustiveSearch-package}
\alias{ExhaustiveSearch}
\alias{ExhaustiveSearch-package}
\title{ExhaustiveSearch: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\details{
A package for fast and scalable exhaustive feature selections of linear and
logistic regression models. Please refer to the individual help pages of the
exhaustive functions or to the github page at
(<https://github.com/RudolfJagdhuber/ExhaustiveSearch/>) for a detailed
introduction on how to use this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RudolfJagdhuber/ExhaustiveSearch}
  \item Report bugs at \url{https://github.com/RudolfJagdhuber/ExhaustiveSearch/issues}
}

}
\keyword{internal}
