% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ExhaustiveSearch}
\alias{print.ExhaustiveSearch}
\title{Print ExhaustiveSearch}
\usage{
\method{print}{ExhaustiveSearch}(x, ...)
}
\arguments{
\item{x}{Object of class 'ExhaustiveSearch'.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
No return value. The function is only called to print results to the
console.
}
\description{
Prints a compact summary of the results of an ExhaustiveSearch object.
}
\seealso{
\code{\link[=ExhaustiveSearch]{ExhaustiveSearch()}}
}
\author{
Rudolf Jagdhuber
}
