% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultFunctions.R
\name{resultTable}
\alias{resultTable}
\title{Overview of the top exhaustive search results}
\usage{
resultTable(ESResult, n = Inf, insertStart = "")
}
\arguments{
\item{ESResult}{a result object from an exhaustive search.}

\item{n}{number of results to be returned. The default (`Inf`) returns every
result available in `ESResult`.}

\item{insertStart}{used for additional spacing when printing. The value of
`insertStart` gets printed in front of every feature combination to
increase the space to the printed performance measure.}
}
\value{
A `data.frame` with two columns. The first one shows the performance
  values and the second shows the decoded feature set collapsed with plus
  signs.
}
\description{
Extract the top `n` results of an exhaustive search and present them as a
`data.frame` object.
}
\details{
The result of an exhaustive search is given by an object of class
ExhaustiveSearch, which is a list of encoded feature combinations and
performance values. This function decodes the feature combinations and
presents them in a `data.frame` together with the respective performance
values
}
\author{
Rudolf Jagdhuber
}
